requirements.txt
absl-py==1.4.0
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
beautifulsoup4==4.9.3
billiard==4.1.0
blessed==1.20.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloud-tpu-client==0.10
colorama==0.4.6
colorlog==6.7.0
confection==0.0.4
cymem==2.0.7
Cython==0.29.32
Django==4.2.2
django-brutebuster2==0.2.0
django-formtools==2.3
django-picklefield==3.1
django-q==1.3.9
docx2pdf==0.1.8
emoji==2.5.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.5.0/es_core_news_sm-3.5.0-py3-none-any.whl
flatbuffers==23.3.3
fst-pso==1.8.1
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google==3.0.0
google-api-core==1.34.0
google-api-python-client==1.8.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
googlesearch-python==1.1.0
grpcio==1.54.0
h5py==3.8.0
httplib2==0.22.0
idna==2.10
inexactsearch==1.0.2
jax==0.4.8
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
keras==2.12.0
Keras-Preprocessing==1.1.2
kombu==5.3.1
langcodes==3.3.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
miniful==0.0.6
ml-dtypes==0.1.0
murmurhash==1.0.9
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathy==0.10.1
preshed==3.0.8
prompt-toolkit==3.0.39
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.4
pyFUME==0.2.25
pyparsing==3.0.9
PyPDF2==2.11.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pywin32==306
PyYAML==6.0
redis==3.5.3
regex==2022.10.31
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
silpa-common==0.3
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
soundex==1.1.3
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spellchecker==0.4
sqlparse==0.4.4
srsly==2.4.5
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.7
threadpoolctl==3.1.0
tika==2.6.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
unicodedata2==15.0.0
uritemplate==3.0.1
urllib3==1.26.14
vine==5.0.0
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.3.0
wrapt==1.14.1
